Stabilize bigint2 feature (#3241)
This moves the stuff marked as "unstable" in `risc0-bigint2` to be
stable.
It also makes bigint2 support no longer require "unstable" inside
`risc0-zkvm` and `risc0-build`

This makes all this code now participate in SemVer checks, which I think
is good so as to prevent downstream compilation issues.

It does not remove the "unstable" feature from `risc0-bigint2`, as this
would break anybody depending on that feature (like all the
precompiles).

I find it rather unfortunate that all the precompiles require the
"unstable" feature, it means in the future if we would want to add new
things marked as "unstable" in `risc0-bigint2`, they would be enabled by
all existing precompiles. I think newer releases of the precompiles
should probably aim to forward the "unstable" feature flag from their
users on to `risc0-bigint2` so the user can decide if they want unstable
features or not themselves.