rzup publish: Add create-artifact, artifact hash and signature verification (#3281)
This adds three new features to the `rzup publish` command
- New `rzup publish create-artifact` which simply creates `.tar.xz`
files
    - Uses the parallel feature of liblzma
- Verification of the SHA256 sums for download payloads
- RSA signing of the S3 JSON payloads
    - The private key is stored in AWS secrets manager

There are some incidental changes here too
- Due to some crate updates / changes requiring a newer compiler verion,
I had to update the nightly version of Rust we use for the docs-rs CI
job, this caused us to need to update some things in random places.
- Even though `rzup` only uses `liblzma` if the `install` or `publish`
features are enabled, it will still complain about a linking conflict
with liblzma between some of the risc0 crates and `rzup`. This meant I
had to use `liblzma` crate in place of the `xz` crate (`liblzma` is just
a newer fork of that crate anyway), and remove the lzma features from
`zip` crate.

---------

Co-authored-by: Victor Graf <victor@risczero.com>