Add support for Groth16 shrink-wrapping to the actor system (#3318)
Adds new `ShrinkWrapRequest` to actor system which is fulfilled by
`shrink_wrap::JobActor`. The `ShrinkWrapRequest` has a `kind` field
which is `ShrinkWrapKind`, the only variant in it right now is
`Groth16`.

There is a new `ShrinkWrap` task type that workers can ask for. We could
split this out more in the future so that workers could subscribe to
only certain kinds of shrink wrapping. The `ShrinkWrap` task always runs
as a GPU task. (the workers don't have support for CPU proving right
now)

Implements the `compress` method of `DefaultProver` when asking for a
groth16 receipt, also add groth16 shrink-wrapping when calling
`prove_with_ctx` and the `receipt_kind` is `Groth16`

Implements the snark API calls in the actor system API.