Extract rzup artifacts out-of-place (#3343)
This should help ensure that components aren't ever half-installed.

The test I add doesn't fail previously because it does try to delete the
directory if the extraction fails, but if rzup is killed it will leave
it half extracted. I can't test the rzup being killed case, but I might
as well backfill the case for regular failed extraction.

With this, I think we should be able to remove the `--force` flag from
the CI invocations