Add execute-only support to the actor-system (#3322)
Implement an execute only option in the actor-system.

This does it by adding a `execute_only` bool to `ProofRequest` and
making the `receipt` in `ProofResult` optional.

I struggled at first trying instead other ways of doing this with its
own "execute-only" job, but the existing bonsai API makes it difficult
to do that.

It implements the `Executor` trait for the `DefaultProver` and also adds
execute-only support to the API part of the actor system. This includes
serializing the journal.

depends on https://github.com/risc0/risc0/pull/3318