Move guest binaries in risc0-povw to allow publishing and add defaults to log-builder prover (#3362)
In https://github.com/risc0/risc0/pull/3330, I added built binaries for
the Log Builder guest. Because I added them to the inner guest
workspace, and then imported them in the library, this fails to build
when pulling from `crates.io`.

This PR moves the pre-built guest binaries. It also sets a default on
the Log Builder prover to use the included bin and ID for easier
instantiation.