Examples and docs for the ApiClient module (#3232)
This PR adds usage examples and documentation to the ApiClient module.
Here is the list of affected methods:
  - `from_env`
  - `prove`
  - `execute`
  - `prove_segment`
  - `lift`
  - `join`
  - `verify`

The usage examples are written in a way to be concise, step by step, and
all-inclusive with the focus being on a pipeline-like proving scheme:
proof = execute -> prove -> lift -> join. Besides examples, a bit of
context have been added to some methods while preserving the original
documentation.

---------

Co-authored-by: rezahsnz <rezahsnz@proton.me>
Co-authored-by: Frank Laub <flaub@risc0.com>