Add a hint for Docker permissioned denied with Groth16 (#3448)
This is a small edit I made while debugging why `bootstrap-groth16` was
not working. One of the issues I ran into was that the process is trying
to run `docker` without sufficient permissions. The error currently just
gives an exit code of `126`, which does indicate a permissions issue but
it took me a while to realize the particular permissions issue. This PR
adds a hint.

I also included two other minor edits:
* Remove `bento/Cargo.toml` from linked projects for the VS Code config.
This file no longer exists.
* Add a `cuda` feature to bootstrap. I decided to do this instead of
using Docker to sidestep the permissions issue.

---------

Co-authored-by: Remi Bernotavicius <bobbobbio@gmail.com>