Add dual-hal implementation for recursion (#3527)
This adds an implementation of the hal for recursion that uses both the
CUDA hal CPU hal and compares the answers.

The use of this HAL is controlled with the existing `dual` feature.

This is very useful for debugging issues with CUDA kernels.

In order to make this work I had to make two changes to other code:
- Add a way to allocate zeroed buffers to the HAL interface. The dual
hal compares buffers without knowing what is initialized or not, so we
need to have it always initialize its buffers
- Allow PartialEq to work to compare `BabyBearElem::INVALID`. Sometimes
the buffer contains this value intentionally and the dual hal wants to
check to make sure this value is present in the places we intend.

---------

Co-authored-by: Jeremy Bruestle <jeremy@risczero.com>
Co-authored-by: Frank Laub <flaub@risc0.com>