Use CUDA streams and don't call `cudaDeviceSynchronize` (#3531)
Put a CUDA stream in the HAL and use it for every kernel. Do all
memcpy's on the stream and synchronize after.

Removing all the calls to `cudaDeviceSynchronize` should help with
parallelism of kernels.

The groth16 code is unchanged, it already uses a sppark stream in a very
similar way, but it would be nice to make it more similar to the other
code at some point.